+++ src/commands/upload/macos.ts
@@ -4,19 +4,19 @@ import * as qq from 'qqjs'
 import aws from '../../aws'
 import {log} from '../../log'
 import * as Tarballs from '../../tarballs'
-import {commitAWSDir} from '../../publish-util'
+import {commitAWSDir} from '../../upload-util'
 
-export default class PublishMacos extends Command {
+export default class UploadMacos extends Command {
   static hidden = true
 
-  static description = 'publish macos installers built with pack:macos'
+  static description = 'upload macos installers built with pack:macos'
 
   static flags = {
     root: flags.string({char: 'r', description: 'path to oclif CLI root', default: '.', required: true}),
   }
 
   async run() {
-    const {flags} = this.parse(PublishMacos)
+    const {flags} = this.parse(UploadMacos)
     const buildConfig = await Tarballs.buildConfig(flags.root)
     const {s3Config, version, config} = buildConfig
     const S3Options = {
@@ -29,6 +29,6 @@ export default class PublishMacos extends Command {
     const key = `${root}${config.bin}.pkg`
     if (await qq.exists(pkg)) await aws.s3.uploadFile(pkg, {...S3Options, CacheControl: 'max-age=86400', Key: key})
 
-    log(`published macos ${version}`)
+    log(`uploaded macos ${version}`)
   }
 }
