+++ src/commands/upload/win.ts
@@ -4,19 +4,19 @@ import * as qq from 'qqjs'
 import aws from '../../aws'
 import {log} from '../../log'
 import * as Tarballs from '../../tarballs'
-import {commitAWSDir} from '../../publish-util'
+import {commitAWSDir} from '../../upload-util'
 
-export default class PublishWin extends Command {
+export default class UploadWin extends Command {
   static hidden = true
 
-  static description = 'publish windows installers built with pack:win'
+  static description = 'upload windows installers built with pack:win'
 
   static flags = {
     root: flags.string({char: 'r', description: 'path to oclif CLI root', default: '.', required: true}),
   }
 
   async run() {
-    const {flags} = this.parse(PublishWin)
+    const {flags} = this.parse(UploadWin)
     const buildConfig = await Tarballs.buildConfig(flags.root)
     const {s3Config, version, config} = buildConfig
     const S3Options = {
@@ -33,6 +33,6 @@ export default class PublishWin extends Command {
     await uploadWin('x64')
     await uploadWin('x86')
 
-    log(`published win ${version}`)
+    log(`uploaded win ${version}`)
   }
 }
